+++ include/CacheBase.h
@@ -46,7 +46,7 @@ namespace openshot {
 	{
 	protected:
 		string cache_type; ///< This is a friendly type name of the derived cache instance
-		long long int max_bytes; ///< This is the max number of bytes to cache (0 = no limit)
+		int64_t max_bytes; ///< This is the max number of bytes to cache (0 = no limit)
 
 		/// Section lock for multiple threads
 	    CriticalSection *cacheCriticalSection;
@@ -58,7 +58,7 @@ namespace openshot {
 
 		/// @brief Constructor that sets the max bytes to cache
 		/// @param max_bytes The maximum bytes to allow in the cache. Once exceeded, the cache will purge the oldest frames.
-		CacheBase(long long int max_bytes);
+		CacheBase(int64_t max_bytes);
 
 		/// @brief Add a Frame to the cache
 		/// @param frame The openshot::Frame object needing to be cached.
@@ -68,33 +68,33 @@ namespace openshot {
 		virtual void Clear() = 0;
 
 		/// Count the frames in the queue
-		virtual long int Count() = 0;
+		virtual int64_t Count() = 0;
 
 		/// @brief Get a frame from the cache
 		/// @param frame_number The frame number of the cached frame
-		virtual std::shared_ptr<Frame> GetFrame(long int frame_number) = 0;
+		virtual std::shared_ptr<Frame> GetFrame(int64_t frame_number) = 0;
 
 		/// Gets the maximum bytes value
-		virtual long long int GetBytes() = 0;
+		virtual int64_t GetBytes() = 0;
 
 		/// Get the smallest frame number
 		virtual std::shared_ptr<Frame> GetSmallestFrame() = 0;
 
 		/// @brief Remove a specific frame
 		/// @param frame_number The frame number of the cached frame
-		virtual void Remove(long int frame_number) = 0;
+		virtual void Remove(int64_t frame_number) = 0;
 
 		/// @brief Remove a range of frames
 		/// @param start_frame_number The starting frame number of the cached frame
 		/// @param end_frame_number The ending frame number of the cached frame
-		virtual void Remove(long int start_frame_number, long int end_frame_number) = 0;
+		virtual void Remove(int64_t start_frame_number, int64_t end_frame_number) = 0;
 
 		/// Gets the maximum bytes value
-		long long int GetMaxBytes() { return max_bytes; };
+		int64_t GetMaxBytes() { return max_bytes; };
 
 		/// @brief Set maximum bytes to a different amount
 		/// @param number_of_bytes The maximum bytes to allow in the cache. Once exceeded, the cache will purge the oldest frames.
-		void SetMaxBytes(long long int number_of_bytes) { max_bytes = number_of_bytes; };
+		void SetMaxBytes(int64_t number_of_bytes) { max_bytes = number_of_bytes; };
 
 		/// @brief Set maximum bytes to a different amount based on a ReaderInfo struct
 		/// @param number_of_frames The maximum number of frames to hold in cache
@@ -102,7 +102,7 @@ namespace openshot {
 		/// @param height The height of the frame's image
 		/// @param sample_rate The sample rate of the frame's audio data
 		/// @param channels The number of audio channels in the frame
-		void SetMaxBytesFromInfo(long int number_of_frames, int width, int height, int sample_rate, int channels);
+		void SetMaxBytesFromInfo(int64_t number_of_frames, int width, int height, int sample_rate, int channels);
 
 		/// Get and Set JSON methods
 		virtual string Json() = 0; ///< Generate JSON string of this object
